requirements.txt
customtkinter>=5.2.0
Pillow>=10.0.0
pyautogui>=0.9.54
pytesseract>=0.3.10
pygame>=2.0.0
pynput>=1.7.6
opencv-python-headless>=4.8.0
numpy>=1.21.0
pywin32>=305
PyYAML>=6.0
